--- controllers/contacts.js
@@ -1,108 +1,0 @@
-const Joi = require("joi");
-const { HttpError, ctrlWrapper } = require("../helpers");
-const { Schema, model } = require("mongoose");
-
-const contactsSchema = new Schema({
-  name: {
-    type: String,
-    required: [true, "Set name for contact"],
-  },
-  email: {
-    type: String,
-  },
-  phone: {
-    type: String,
-  },
-  favorite: {
-    type: Boolean,
-    default: false,
-  },
-});
-
-contactsSchema.post("save", (error, data, next) => {
-  error.status = 400;
-  next();
-});
-
-const Contact = model("contact", contactsSchema);
-
-const addSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-  favorite: Joi.boolean(),
-});
-
-const updateFavoriteSchema = Joi.object({
-  favorite: Joi.boolean().required(),
-});
-
-const getAllContacts = async (req, res) => {
-  const result = await Contact.find();
-  res.json(result);
-};
-
-const getContactById = async (req, res) => {
-  const { contactId } = req.params;
-  const contactById = await Contact.findById(contactId);
-  if (!contactById) {
-    throw HttpError(404, "Not found");
-  }
-  res.json(contactById);
-};
-
-const addContact = async (req, res) => {
-  const newContact = await Contact.create(req.body);
-  res.status(201).json(newContact);
-};
-
-const deleteContact = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndRemove(contactId);
-  if (!result) {
-    throw HttpError(404, "Not found");
-  }
-  res.status(200).json({
-    message: "Contact deleted",
-  });
-};
-
-const updateContact = async (req, res) => {
-  const { contactId } = req.params;
-  const { error } = addSchema.validate(req.body);
-  if (error) {
-    throw HttpError(400, error.message);
-  }
-  const updatedContact = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
-  if (!updatedContact) {
-    throw HttpError(404, "Not found");
-  }
-  res.status(200).json(updatedContact);
-};
-
-const updateStatusContact = async (req, res) => {
-  const { contactId } = req.params;
-  if (!req.body) {
-    throw HttpError(400, "missing field favorite");
-  }
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
-  if (!result) {
-    throw HttpError(404, "Not found");
-  }
-  res.status(200).json(result);
-};
-
-module.exports = {
-  getAllContacts: ctrlWrapper(getAllContacts),
-  getContactById: ctrlWrapper(getContactById),
-  addContact: ctrlWrapper(addContact),
-  deleteContact: ctrlWrapper(deleteContact),
-  updateContact: ctrlWrapper(updateContact),
-  updateStatusContact: ctrlWrapper(updateStatusContact),
-  addSchema,
-  updateFavoriteSchema,
-};
