+++ controllers/contacts/deleteContact.js
@@ -1,0 +1,15 @@
+const { Contact } = require("../../models/contacts");
+const { HttpError } = require("../../helpers");
+
+const deleteContact = async (req, res) => {
+  const { contactId } = req.params;
+  const result = await Contact.findByIdAndRemove(contactId);
+  if (!result) {
+    throw HttpError(404, "Not found");
+  }
+  res.status(200).json({
+    message: "Contact deleted",
+  });
+};
+
+module.exports = deleteContact;
