+++ models/contacts.js
@@ -1,1 +1,43 @@
-/* сюда нужно перенести Contact.schema и Joi.schema 44:31 */
+const Joi = require("joi");
+const { Schema, model } = require("mongoose");
+
+const contactsSchema = new Schema({
+  name: {
+    type: String,
+    required: [true, "Set name for contact"],
+  },
+  email: {
+    type: String,
+  },
+  phone: {
+    type: String,
+  },
+  favorite: {
+    type: Boolean,
+    default: false,
+  },
+});
+
+contactsSchema.post("save", (error, data, next) => {
+  error.status = 400;
+  next();
+});
+
+const Contact = model("contact", contactsSchema);
+
+const addSchema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().required(),
+  phone: Joi.string().required(),
+  favorite: Joi.boolean(),
+});
+
+const updateFavoriteSchema = Joi.object({
+  favorite: Joi.boolean().required(),
+});
+
+module.exports = {
+  addSchema,
+  updateFavoriteSchema,
+  Contact,
+};
