+++ models/index.js
@@ -1,0 +1,7 @@
+const { addSchema, updateFavoriteSchema, Contact } = require("./contacts");
+
+module.exports = {
+  addSchema,
+  updateFavoriteSchema,
+  Contact,
+};
