+++ routes/api/contacts.js
@@ -1,10 +1,8 @@
 const express = require("express");
 const router = express.Router();
-const ctrl = require("../../controllers/contacts");
-const {
-  addSchema,
-  updateFavoriteSchema,
-} = require("../../controllers/contacts");
+const ctrl = require("../../controllers/contacts/index");
+console.log("🚀 ~ ctrl:", ctrl);
+const { addSchema, updateFavoriteSchema } = require("../../models/contacts");
 const { validateBody, isValidId } = require("../../middlewares");
 
 router.get("/", ctrl.getAllContacts);
